Move `WrongNetworkSelectedDialog` to $shared
