+++ app/src/marketplace/containers/EditProductPage/DeployDataUnionModal.jsx
@@ -7,7 +7,7 @@ import { type Product } from '$mp/flowtype/product-types'
 import GuidedDeployDataUnionDialog from '$mp/components/Modal/GuidedDeployDataUnionDialog'
 import ConfirmDeployDataUnionDialog from '$mp/components/Modal/ConfirmDeployDataUnionDialog'
 import DeployingDataUnionDialog from '$mp/components/Modal/DeployingDataUnionDialog'
-import WrongNetworkSelectedDialog from '$mp/components/Modal/WrongNetworkSelectedDialog'
+import WrongNetworkSelectedDialog from '$shared/components/WrongNetworkSelectedDialog'
 import ErrorDialog from '$mp/components/Modal/ErrorDialog'
 import { isLocalStorageAvailable } from '$shared/utils/storage'
 import { deployDataUnion } from '$mp/modules/dataUnion/services'
