+++ app/src/marketplace/containers/EditProductPage/PublishModal.jsx
@@ -14,7 +14,7 @@ import ReadyToPublishDialog from '$mp/components/Modal/ReadyToPublishDialog'
 import PublishTransactionProgress from '$mp/components/Modal/PublishTransactionProgress'
 import PublishComplete from '$mp/components/Modal/PublishComplete'
 import PublishError from '$mp/components/Modal/PublishError'
-import WrongNetworkSelectedDialog from '$mp/components/Modal/WrongNetworkSelectedDialog'
+import WrongNetworkSelectedDialog from '$shared/components/WrongNetworkSelectedDialog'
 import Web3ErrorDialog from '$shared/components/Web3ErrorDialog'
 import useWeb3Status from '$shared/hooks/useWeb3Status'
 import UnlockWalletDialog from '$shared/components/Web3ErrorDialog/UnlockWalletDialog'
