+++ app/src/marketplace/containers/ProductPage/PurchaseModal.jsx
@@ -23,7 +23,7 @@ import PurchaseError from '$mp/components/Modal/PurchaseError'
 import ErrorDialog from '$mp/components/Modal/ErrorDialog'
 import NoBalanceDialog from '$mp/components/Modal/NoBalanceDialog'
 import ChooseAccessPeriodDialog from '$mp/components/Modal/ChooseAccessPeriodDialog'
-import WrongNetworkSelectedDialog from '$mp/components/Modal/WrongNetworkSelectedDialog'
+import WrongNetworkSelectedDialog from '$shared/components/WrongNetworkSelectedDialog'
 import useIsMounted from '$shared/hooks/useIsMounted'
 import type { Ref, UseStateTuple } from '$shared/flowtype/common-types'
 import Web3ErrorDialog from '$shared/components/Web3ErrorDialog'
